+++ DataConnectors/microsoft-sentinel-logstash-output-plugin/CHANGELOG.md
@@ -1,2 +1,6 @@
 ## 1.0.0
 * Initial release for output plugin for logstash to Microsoft Sentinel. This is done with the Log Analytics DCR based API.
+
+## 1.0.2
+* Upgrade the rest-client dependency minimum version to 2.1.0
+* Allow setting different proxy values for api connections.